Adding session management.

+++ src/main/java/com/base/test/controller/AbstractRestController.java
@@ -1,7 +1,9 @@
 package com.base.test.controller;
 
-import java.util.List;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpSession;
 
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -12,18 +14,32 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 
 import com.base.test.Services.ServiceInterface;
+import com.base.test.model.User;
 
 public abstract class AbstractRestController<T> implements RestControllerInterface<T> {
 
+	@Autowired
+	private ServiceInterface<User> userService;
+
 	public abstract ServiceInterface<T> getEntityService();
 
+	@Override
 	@GetMapping("/getAll")
-	public List<T> getEntity() {
-		return getEntityService().getAll();
+	public ResponseEntity getEntity(HttpServletRequest request) {
+		String userName = checkUser(request);
+		if (userName == null) {
+			return new ResponseEntity("User is not loged in ", HttpStatus.UNAUTHORIZED);
+		}
+		return new ResponseEntity(getEntityService().getAll(), HttpStatus.OK);
 	}
 
+	@Override
 	@GetMapping("/get/{id}")
-	public ResponseEntity getEntityByID(@PathVariable("id") Long id) {
+	public ResponseEntity getEntityByID(@PathVariable("id") Long id, HttpServletRequest request) {
+		String userName = checkUser(request);
+		if (userName == null) {
+			return new ResponseEntity("User is not loged in ", HttpStatus.UNAUTHORIZED);
+		}
 		T table = getEntityService().findByID(id);
 		if (table == null) {
 			return new ResponseEntity("No Entity found for ID " + id, HttpStatus.NOT_FOUND);
@@ -31,22 +47,37 @@ public abstract class AbstractRestController<T> implements RestControllerInterfa
 		return new ResponseEntity(table, HttpStatus.OK);
 	}
 
+	@Override
 	@PostMapping(value = "/create")
-	public ResponseEntity createEntity(@RequestBody T tables) {
+	public ResponseEntity createEntity(@RequestBody T tables, HttpServletRequest request) {
+		String userName = checkUser(request);
+		if (userName == null) {
+			return new ResponseEntity("User is not loged in ", HttpStatus.UNAUTHORIZED);
+		}
 		getEntityService().create(tables);
 		return new ResponseEntity(tables, HttpStatus.OK);
 	}
 
+	@Override
 	@DeleteMapping("/delete/{id}")
-	public ResponseEntity deleteEntity(@PathVariable Long id) {
+	public ResponseEntity deleteEntity(@PathVariable Long id, HttpServletRequest request) {
+		String userName = checkUser(request);
+		if (userName == null) {
+			return new ResponseEntity("User is not loged in ", HttpStatus.UNAUTHORIZED);
+		}
 		if (null == getEntityService().delete(id)) {
 			return new ResponseEntity("No Entity found for ID " + id, HttpStatus.NOT_FOUND);
 		}
 		return new ResponseEntity(id, HttpStatus.OK);
 	}
 
+	@Override
 	@PutMapping("/update/{id}")
-	public ResponseEntity updateEntity(@PathVariable Long id, @RequestBody T tables) {
+	public ResponseEntity updateEntity(@PathVariable Long id, @RequestBody T tables, HttpServletRequest request) {
+		String userName = checkUser(request);
+		if (userName == null) {
+			return new ResponseEntity("User is not loged in ", HttpStatus.UNAUTHORIZED);
+		}
 		tables = getEntityService().update(id, tables);
 		if (null == tables) {
 			return new ResponseEntity("No Entity found for ID " + id, HttpStatus.NOT_FOUND);
@@ -54,4 +85,12 @@ public abstract class AbstractRestController<T> implements RestControllerInterfa
 		return new ResponseEntity(tables, HttpStatus.OK);
 	}
 
+	public String checkUser(HttpServletRequest request) {
+		HttpSession session = request.getSession();
+		String userName = (String) session.getAttribute("UserName");
+		if (userName == null)
+			return null;
+		User user = userService.getByName(userName);
+		return user != null ? user.getUserName() : null;
+	}
 }
